rename class name, variable name.

+++ src/main/java/ru/job4j/converter/Converter.java
@@ -1,19 +1,19 @@
 package ru.job4j.converter;
 
 public class Converter {
-    public static int rubleToEuro(int value) {
+    public static float rubleToEuro(float value) {
         return value / 70;
     }
 
-    public static int rubleToDollar(int value) {
+    public static float rubleToDollar(float value) {
         return value / 60;
     }
 
     public static void main(String[] args) {
-        int euro = Converter.rubleToEuro(140);
+        float euro = Converter.rubleToEuro(140);
         System.out.println("140 rubles are " + euro + " euro.");
 
-        int dollar = Converter.rubleToDollar(140);
+        float dollar = Converter.rubleToDollar(140);
         System.out.println("140 rubles are " + dollar + " dollar.");
     }
 }
